fix extenSions
